projects/arxiv-ai-hardware-daily-digest/requirements.txt
# Arxiv AI Hardware Daily Digest - 依赖包

# 核心依赖
python>=3.10

# 数据抓取和处理
arxiv==2.1.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# 文本处理和分析
nltk>=3.8.0
spacy>=3.5.0
# 下载spacy模型: python -m spacy download en_core_web_sm

# 飞书集成
feishu-sdk>=1.0.0  # 或使用 requests 直接调用API

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 日期时间处理
python-dateutil>=2.8.0
pytz>=2023.0

# 日志和监控
structlog>=23.0.0
prometheus-client>=0.17.0

# 测试框架
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0

# 异步支持（可选）
aiohttp>=3.8.0
asyncio>=3.4.3